+++ test/parallel/test-tick-processor-cpp-core.js
@@ -1,0 +1,35 @@
+'use strict';
+const common = require('../common');
+const fs = require('fs');
+const cp = require('child_process');
+const path = require('path');
+
+// TODO(mhdawson) Currently the test-tick-processor functionality in V8
+// depends on addresses being smaller than a full 64 bits.  Aix supports
+// the full 64 bits and the result is that it does not process the
+// addresses correctly and runs out of memory
+// Disabling until we get a fix upstreamed into V8
+if (common.isAix) {
+  common.skip('Aix address range too big for scripts.');
+  return;
+}
+
+const base = require(path.join(common.fixturesDir, 'tick-processor-base.js'));
+
+if (common.isWindows ||
+    common.isSunOS ||
+    common.isAix ||
+    common.isLinuxPPCBE ||
+    common.isFreeBSD) {
+  common.skip('C++ symbols are not mapped for this os.');
+  return;
+}
+
+base.runTest({
+  pattern: /RunInDebugContext/,
+  code: `function f() {
+           require(\'vm\').runInDebugContext(\'Debug\');
+           setImmediate(function() { f(); });
+         };
+         f();`
+});
